% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base32_handling.R
\name{base32_encode}
\alias{base32_encode}
\title{encode integer array describing the deck list
into base32 representation of deck code}
\usage{
base32_encode(int_decklist)
}
\description{
encode integer array describing the deck list
into base32 representation of deck code
}
