% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decklist_handling.R
\name{df_to_simple_decklist}
\alias{df_to_simple_decklist}
\title{parse dataframe decklist into simple format}
\usage{
df_to_simple_decklist(decklist_df)
}
\arguments{
\item{decklist_df}{dataframe with character column \emph{cardcode} and numeric column \emph{count}}
}
\value{
character vector of decklist with format '\{count\}:\{cardcode\}'
}
\description{
Intended for compatibility of the original decklist format and the more R native use
of dataframes
}
