% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decklist_handling.R
\name{gen_int_decklist}
\alias{gen_int_decklist}
\title{translate decklist with cards and their count
into an integer representation following the rules
laid out in https://github.com/RiotGames/LoRDeckCodes/blob/master/README.md}
\usage{
gen_int_decklist(decklist, faction_lut = get_faction_lut())
}
\description{
translate decklist with cards and their count
into an integer representation following the rules
laid out in https://github.com/RiotGames/LoRDeckCodes/blob/master/README.md
}
