% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel_functions.R
\name{get_code_from_decklist}
\alias{get_code_from_decklist}
\title{Generate a shareable deck code from a decklist}
\usage{
get_code_from_decklist(decklist)
}
\arguments{
\item{decklist}{dataframe with character column \emph{cardcode} and numeric column \emph{count}}
}
\value{
character deck code in base 32 (RFC 4648)
}
\description{
Deck lists are stored and shared in an \emph{RFC 4648} base-32 code. This top-level
function takes the cards and their count , and encodes them according to the rules
(\href{https://github.com/RiotGames/LoRDeckCodes}{laid out here})
}
\examples{
#minimalist example (obvisouly incomplete deck)
min_deck <- data.frame(
  "cardcode" = c("01PZ008", "01PZ040"),
  "count" = c(3, 3)
)

get_code_from_decklist(min_deck)

# returns "CEAQEAIEBAUAAAA"
}
