% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decklist_handling.R
\name{simple_decklist_to_df}
\alias{simple_decklist_to_df}
\title{parse simple decklist format into dataframe}
\usage{
simple_decklist_to_df(decklist)
}
\arguments{
\item{decklist}{character vector of decklist with format '\{count\}:\{cardcode\}'}
}
\value{
dataframe with character column \emph{cardcode} and numeric column \emph{count}
}
\description{
Intended for compatibility of the original decklist format and the more R native use
of dataframes
}
\examples{
lordecks:::simple_decklist_to_df(c("3:01PZ008", "3:01PZ040"))
# returns  :
#   count cardcode
# 1     3  01PZ008
# 2     3  01PZ040
}
