% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varint_handling.R
\name{varint_bytes_to_binary}
\alias{varint_bytes_to_binary}
\title{translate varint bytes into binary representation
of encoded integers}
\usage{
varint_bytes_to_binary(bytearray)
}
\description{
translate varint bytes into binary representation
of encoded integers
}
