% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base32_handling.R
\name{base32_decode}
\alias{base32_decode}
\title{decode base32 representation of deck code
into integer array describing the deck list}
\usage{
base32_decode(code)
}
\description{
decode base32 representation of deck code
into integer array describing the deck list
}
