% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decklist_handling.R
\name{parse_decklist}
\alias{parse_decklist}
\title{parse integer representation of decklist
returns dataframe with cardcode, count, faction, set and card number}
\usage{
parse_decklist(cardinfo, faction_lut = get_faction_lut())
}
\description{
parse integer representation of decklist
returns dataframe with cardcode, count, faction, set and card number
}
